module: use nautilus_extension::NautilusModule
