+++ Cargo.toml
@@ -13,6 +13,7 @@ gio-sys = "0.3.1"
 glib-sys = "0.3.1"
 gobject-sys = "0.3.1"
 libc = "0.2"
+nautilus-extension = { path = "../nautilus-extension-rs/nautilus-extension" }
 nautilus-extension-sys = { path = "../nautilus-extension-rs/nautilus-extension-sys" }
 url = "1.2.0"
