+++ src/tmsu_nautilus.rs
@@ -1,0 +1,43 @@
+use glib_ffi::GType;
+use gobject_ffi::{GInterfaceInfo, GTypeModule};
+use gobject_ffi::g_type_module_add_interface;
+use libc::c_int;
+use nautilus_extension::NautilusModule;
+use nautilus_ffi::{nautilus_column_provider_get_type, nautilus_info_provider_get_type, nautilus_menu_provider_get_type};
+use std::ptr;
+
+use column_provider;
+use info_provider;
+use menu_provider;
+
+nautilus_module!(register_type);
+
+fn register_type(module: *mut GTypeModule) -> GType {
+    let module_type =
+        NautilusModule::new(module, "TmsuNautilusExtension")
+        .register();
+
+    let column_provider_iface_info = GInterfaceInfo {
+        interface_init: Some(column_provider::tmsu_extension_column_provider_iface_init),
+        interface_finalize: None,
+        interface_data: ptr::null_mut()
+    };
+    let info_provider_iface_info = GInterfaceInfo {
+        interface_init: Some(info_provider::tmsu_extension_info_provider_iface_init),
+        interface_finalize: None,
+        interface_data: ptr::null_mut()
+    };
+    let menu_provider_iface_info = GInterfaceInfo {
+        interface_init: Some(menu_provider::tmsu_extension_menu_provider_iface_init),
+        interface_finalize: None,
+        interface_data: ptr::null_mut()
+    };
+
+    unsafe {
+        g_type_module_add_interface(module, module_type, nautilus_column_provider_get_type(), &column_provider_iface_info);
+        g_type_module_add_interface(module, module_type, nautilus_info_provider_get_type(), &info_provider_iface_info);
+        g_type_module_add_interface(module, module_type, nautilus_menu_provider_get_type(), &menu_provider_iface_info);
+    }
+
+    module_type
+}
